refactor out LuaBridge and ScriptBridge
