+++ CMakeLists.txt
@@ -32,7 +32,7 @@ FetchContent_Declare(
 FetchContent_Declare(
   ionic
   GIT_REPOSITORY https://github.com/leethomason/ionic.git
-  GIT_TAG        74cc38510f5054d1475530b591351630adacd19a
+  GIT_TAG        9c58d79eeb6820b2bc05740eeddf368985d7b031
 )
 FetchContent_Declare(
   plog
