+++ drivers/console/consoleboard.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <string>
+
+void ConsoleBoardDriver(const std::string& gameDir);
